requirements.txt
PsychoPy==2022.2.5
pyzmq==26.2.0
msgpack==1.1.0
numpy==1.23.5